# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ISI-Eco-Report API",
        "description": "API de signalement et de gestion des déchets sauvages. Permet l'authentification des utilisateurs (Citoyens, Agents et Administrateurs), le signalement géolocalisé avec photo, l'affectation à des brigades de collecte, le suivi des interventions terrain et l'attribution automatique de points de fidélité.",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription citoyenne",
                "description": "Permet aux nouveaux citoyens de créer un compte. Rôle par défaut : 'citizen'.",
                "operationId": "f18a0c4b85d7508bf5f73ad13af12198",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Sow"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Mamadou"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mamadou@gmail.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "775283814"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Fann Résidence, Dakar"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Citoyen créé avec succès"
                    },
                    "422": {
                        "description": "Données de validation invalides"
                    }
                },
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion de l'utilisateur",
                "description": "Permet aux Citoyens, Agents et Admins de se connecter et d'obtenir un jeton d'accès.",
                "operationId": "c6c9874077b7f3ebeda92067d8d9fa9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@isieco.sn"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie avec token"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                },
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "description": "Révoque le jeton d'accès actuel de l'utilisateur connecté.",
                "operationId": "767c40d11a53a4b9ef07c97c6353be52",
                "responses": {
                    "204": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Demander un lien de réinitialisation de mot de passe",
                "description": "Envoie un e-mail contenant le lien de réinitialisation si l'adresse e-mail existe.",
                "operationId": "b1ff7c459a9fec2ac05541848a768b83",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mamadou@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien envoyé"
                    }
                },
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Réinitialisation du mot de passe",
                "description": "Réinitialise le mot de passe de l'utilisateur à l'aide du jeton reçu par e-mail.",
                "operationId": "92f72120af6836efb02f39cde32189ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "token_recu"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mamadou@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe modifié avec succès"
                    }
                },
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
        "/equipes": {
            "get": {
                "tags": [
                    "Équipes"
                ],
                "summary": "Lister les équipes",
                "description": "Accès : Admin et Agent de collecte.",
                "operationId": "d1b6aaa538683f55cf8bd02c714391e8",
                "responses": {
                    "200": {
                        "description": "Liste des équipes récupérée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
        "/signalements": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Lister les signalements",
                "description": "Accès : Tout utilisateur. Les citoyens voient uniquement leurs propres signalements. Les admins/agents voient tous les signalements.",
                "operationId": "49d2c9d81862854ebb2cdbc14914cc45",
                "responses": {
                    "200": {
                        "description": "Liste des signalements récupérée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
        "/affectations": {
            "get": {
                "tags": [
                    "Affectations"
                ],
                "summary": "Lister les affectations",
                "description": "Accès : Admin et Agent uniquement.",
                "operationId": "65be21fd7b38a706a1dc7d863b94aaf9",
                "responses": {
                    "200": {
                        "description": "Liste des affectations récupérée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
        "/interventions": {
            "get": {
                "tags": [
                    "Interventions"
                ],
                "summary": "Lister les interventions",
                "description": "Accès : Admin et Agent uniquement.",
                "operationId": "ce1611ef2aa04d7e07b334f39659ec80",
                "responses": {
                    "200": {
                        "description": "Liste des interventions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api",
                        "description": "Serveur API local"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insérez votre jeton d'accès brut obtenu lors de la connexion (sans le préfixe 'Bearer ').",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}